+++ Server/src/main/java/com/codestates/comment/service/CommentService.java
@@ -5,9 +5,11 @@ import com.codestates.comment.entity.Comment;
 import com.codestates.comment.repository.CommentRepository;
 import com.codestates.exception.BusinessLogicException;
 import com.codestates.exception.ExceptionCode;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Service
@@ -20,11 +22,18 @@ public class CommentService {
         this.answerRepository = answerRepository;
     }
     public Comment createComment(Comment comment) {
+        Long memberIdFromToken = getMemberIdFromToken();
+        comment.getMember().setMemberId(memberIdFromToken);
         return commentRepository.save(comment);
     }
 
     public Comment updateComment(Comment comment) {
         Comment findComment = findVerifiedComment(comment.getCommentId());
+        Long memberIdFromToken = getMemberIdFromToken();
+
+        if(comment.getMember().getMemberId() != memberIdFromToken) {
+            throw new BusinessLogicException(ExceptionCode.CANNOT_CHANGE_COMMENT);
+        }
 
         Optional.ofNullable(comment.getContent())
                 .ifPresent(content -> findComment.setContent(content));
@@ -34,10 +43,17 @@ public class CommentService {
 
     public void deleteComment(long commentId) {
         Comment findComment = findVerifiedComment(commentId);
+        Long memberIdFromToken = getMemberIdFromToken();
+        if(findComment.getMember().getMemberId() != memberIdFromToken) {
+            throw new BusinessLogicException(ExceptionCode.CANNOT_DELETE_COMMENT);
+        }
 
         commentRepository.delete(findComment);
     }
 
+    public List findComments(long answerId) {
+        return commentRepository.findByAnswer_AnswerId(answerId);
+    }
 
 
     private Comment findVerifiedComment(long commentId) {
@@ -46,5 +62,19 @@ public class CommentService {
                 new BusinessLogicException(ExceptionCode.ANSWER_NOT_FOUND));
 
         return findComment;
+    }
+
+    private Long getMemberIdFromToken() {
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        //principal 헤더가 비어있을 수 있음 (로그인 안된 상태)
+        if(!(principal instanceof Map)) {
+            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_LOGIN);
+        }
+
+        Map findPrincipal = (Map) principal;
+        Long memberId = (Long) findPrincipal.get("memberId");
+
+        return memberId;
     }
 }
